+++ src/storm-gspn/builder/ExplicitGspnModelBuilder.cpp
@@ -5,7 +5,7 @@
 //#include "storm/utility/macros.h"
 //#include "storm/exceptions/NotImplementedException.h"
 //#include "storm/storage/expressions/ExpressionManager.h"
-//#include "storm/parser/FormulaParser.h"
+//#include "storm-parsers/parser/FormulaParser.h"
 //#include "storm/storage/expressions/ExpressionEvaluator.h"
 //
 //namespace storm {