+++ src/test/storm/storage/JaniModelTest.cpp
@@ -1,6 +1,6 @@
 #include "gtest/gtest.h"
 #include "storm-config.h"
-#include "storm/parser/PrismParser.h"
+#include "storm-parsers/parser/PrismParser.h"
 
 #include "storm/utility/solver.h"